move providers to the "providers" folder
