+++ apps/remix-ide/src/app.js
@@ -27,13 +27,13 @@ import { StoragePlugin } from './app/plugins/storage'
 import { Layout } from './app/panels/layout'
 import { NotificationPlugin } from './app/plugins/notification'
 import { Blockchain } from './blockchain/blockchain.js'
-import { HardhatProvider } from './app/tabs/hardhat-provider'
-import { GanacheProvider } from './app/tabs/ganache-provider'
-import { FoundryProvider } from './app/tabs/foundry-provider'
-import { ExternalHttpProvider } from './app/tabs/external-http-provider'
-import { BasicInjectedProvider } from './app/tabs/basic-injected-provider'
-import { Injected0ptimismProvider } from './app/tabs/injected-optimism-provider'
-import { InjectedArbitrumOneProvider } from './app/tabs/injected-arbitrum-one-provider'
+import { HardhatProvider } from './app/providers/hardhat-provider'
+import { GanacheProvider } from './app/providers/ganache-provider'
+import { FoundryProvider } from './app/providers/foundry-provider'
+import { ExternalHttpProvider } from './app/providers/external-http-provider'
+import { BasicInjectedProvider } from './app/providers/basic-injected-provider'
+import { Injected0ptimismProvider } from './app/providers/injected-optimism-provider'
+import { InjectedArbitrumOneProvider } from './app/providers/injected-arbitrum-one-provider'
 import { FileDecorator } from './app/plugins/file-decorator'
 import { CodeFormat } from './app/plugins/code-format'
 import { SolidityUmlGen } from './app/plugins/solidity-umlgen'
