+++ apps/remix-ide/src/app/providers/injected-arbitrum-one-provider.tsx
